update XType API to match type used in JNA netcdf mappings
git-svn-id: https://cida-svn.er.usgs.gov/repos/dev/usgs/netcdf/netcdf-jna/trunk@11321 622dc1fb-2224-0410-ab72-dc85a3975e1a

+++ src/main/java/edu/unidata/ucar/netcdf/jna/NCUtil.java
@@ -151,7 +151,7 @@ public class NCUtil {
 			return sizeBytes;
 		}
 
-		public long getSizeBytes(int count) {
+		public long getSizeBytes(long count) {
 			if (count < 0) {
 				throw new IllegalArgumentException("count parameter < 0");
 			}
